pyproject.toml
[project]
name = "gis-city"
version = "0.1.0"
description = "POI data import tool for PostGIS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sqlalchemy>=2.0.0",
    "geoalchemy2>=0.15.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "geopandas>=1.0.0",
    "shapely>=2.0.0",
]
